🔨 Refactoring: post/PassPosting 로직 분리
