+++ src/pages/posting/Posting.tsx
@@ -1,18 +1,14 @@
 import { useLocation } from 'react-router-dom';
 import { useRecoilValue } from 'recoil';
 
-import postCreateNewPost from '@apis/posting';
 import { userState } from '@/states/userState';
-import { makeFormData } from './utils';
-import { NewPost } from './components';
-import { Link } from '@components/Link';
+import { NewPost, PassPosting } from './components';
 import { LandingMain } from '@pages/landing/style';
 import { POSTING_DESCRIPTION } from '@pages/posting/constants';
 import {
   StyledPosting,
   ContentContainer,
   StyledDescription,
-  StyledPassPosting
 } from './Posting.style';
 
 const Posting = () => {
@@ -22,27 +18,15 @@ const Posting = () => {
   const { token } = useRecoilValue(userState);
   const customToken = `bearer ${token}`;
 
-  const { HEADER, PASS_POSTING } = POSTING_DESCRIPTION;
-
-  const handleClickPassPost = () => {
-    const formData = makeFormData('', channelId);
-    postCreateNewPost(customToken, formData);
-  };
+  const { HEADER } = POSTING_DESCRIPTION;
 
   return (
     <LandingMain>
       <StyledPosting>
         <ContentContainer>
           <StyledDescription>{HEADER}</StyledDescription>
-          <NewPost channelId={channelId} />
-          <StyledPassPosting onClick={handleClickPassPost}>
-            <Link
-              pageLink={'/posts'}
-              size={14}
-              color={'white'}>
-              {PASS_POSTING}
-            </Link>
-          </StyledPassPosting>
+          <NewPost channelId={channelId} customToken={customToken} />
+          <PassPosting channelId={channelId} customToken={customToken} />
         </ContentContainer>
       </StyledPosting>
     </LandingMain>
