+++ src/pages/posting/components/PassPosting.style.ts
@@ -1,0 +1,6 @@
+import styled from '@emotion/styled';
+
+export const StyledPassPosting = styled.div`
+  ${({ theme }) => theme.style.flexCenter};
+  height: 50px;
+`;
